[멤버] 테스트 코드 작성

+++ src/main/java/poomasi/domain/member/_profile/dto/response/MemberProfileResponse.java
@@ -9,8 +9,8 @@ public record MemberProfileResponse(
         String phoneNumber,
         String defaultAddress,
         String addressDetail,
-        Long coordinateX,
-        Long coordinateY,
+        Double coordinateX,
+        Double coordinateY,
         boolean isBanned,
         LocalDateTime createdAt,
         Image profileImage){
